Change balance return type from f32 to f64
Update the `balance` method to return `f64` instead of `f32`. This ensures higher precision for the balance calculation, accommodating larger values accurately.

+++ src/database/account.rs
@@ -57,8 +57,8 @@ impl Account {
         Ok(pubkey)
     }
 
-    pub fn balance(&self) -> f32 {
-        lamports_to_sol(self.balance.unwrap_or_else(|| 0u64)) as f32
+    pub fn balance(&self) -> f64 {
+        lamports_to_sol(self.balance.unwrap_or_else(|| 0u64))
     }
 
     pub fn account_keypair(&self) -> Result<Keypair, Box <dyn std::error::Error>> {
